sports_prediction_ai/requirements.txt
requests
pandas
scikit-learn
xgboost
joblib
matplotlib
seaborn
numpy
PyYAML
kaggle
pytest
pytest-mock
coverage
